💥 Equipmentの仕様変更(Game内のみに仕様を制限)
